refactor(weak-lensing): split internals into focused submodules
Extract weak-lensing arguments, source logic, and factory classes from
the monolithic private module into dedicated private submodules. Keep
systematics in a renamed _systematics module and update package
re-exports to preserve the public firecrown.likelihood.weak_lensing
API, including WeakLensingSystematicFactory.

Update internal and test imports to target the package namespace
instead of the old private _weak_lensing module, preserving behavior
while improving module boundaries and maintainability.
