+++ firecrown/app/sacc/_transform.py
@@ -7,8 +7,8 @@ from pathlib import Path
 from typing import Annotated, assert_never
 
 import sacc
-from sacc.utils import detect_sacc_file_type
 import typer
+from sacc.utils import detect_sacc_file_type
 
 from firecrown.metadata_functions import (
     extract_all_harmonic_metadata_indices,
