+++ firecrown/likelihood/_two_point.py
@@ -23,8 +23,8 @@ import firecrown.generators as gen
 import firecrown.metadata_types as mdt
 from firecrown.data_types import DataVector, TheoryVector, TwoPointMeasurement
 from firecrown.likelihood._cmb import CMBConvergence, CMBConvergenceFactory
-from firecrown.likelihood._weak_lensing import WeakLensing, WeakLensingFactory
 from firecrown.likelihood.number_counts import NumberCounts, NumberCountsFactory
+from firecrown.likelihood.weak_lensing import WeakLensing, WeakLensingFactory
 from firecrown.likelihood_base import Source, Statistic, Tracer
 from firecrown.metadata_functions import (
     TwoPointHarmonicIndex,
