+++ firecrown/likelihood/weak_lensing/_systematics.py
@@ -1,55 +1,23 @@
-"""Weak lensing source and systematics."""
+"""Systematic classes for weak lensing sources."""
 
 from __future__ import annotations
 
 from abc import abstractmethod
-from collections.abc import Sequence
-from dataclasses import dataclass, replace
-from typing import Annotated, Literal, final
+from dataclasses import replace
 
 import numpy as np
-import numpy.typing as npt
 import pyccl
 import pyccl.nl_pt
-import sacc
-from pydantic import BaseModel, ConfigDict, Field, PrivateAttr
 
+from firecrown.likelihood.weak_lensing._args import WeakLensingArgs
 from firecrown.likelihood_base import (
-    PhotoZShiftandStretchFactory,
-    PhotoZShiftFactory,
-    SourceGalaxy,
-    SourceGalaxyArgs,
     SourceGalaxyPhotoZShift,
     SourceGalaxyPhotoZShiftandStretch,
     SourceGalaxySelectField,
     SourceGalaxySystematic,
-    Tracer,
 )
-from firecrown.metadata_types import InferredGalaxyZDist, TypeSource
 from firecrown.modeling_tools import ModelingTools
-from firecrown.updatable import ParamsMap, register_new_updatable_parameter
-
-
-@dataclass(frozen=True)
-class WeakLensingArgs(SourceGalaxyArgs):
-    """Class for weak lensing tracer builder argument."""
-
-    ia_bias: None | tuple[npt.NDArray[np.float64], npt.NDArray[np.float64]] = None
-
-    ia_amplitude: None | np.float64 = None
-    ia_mass_scaling: None | np.float64 = None
-    red_fraction: None | np.float64 = None
-    log10_average_halo_mass: None | np.float64 = None
-
-    has_pt: bool = False
-    has_hm: bool = False
-
-    ia_pt_c_1: None | tuple[npt.NDArray[np.float64], npt.NDArray[np.float64]] = None
-    ia_pt_c_d: None | tuple[npt.NDArray[np.float64], npt.NDArray[np.float64]] = None
-    ia_pt_c_2: None | tuple[npt.NDArray[np.float64], npt.NDArray[np.float64]] = None
-
-    ia_a_1h: None | npt.NDArray[np.float64] = None
-    ia_a_2h: None | npt.NDArray[np.float64] = None
+from firecrown.updatable import register_new_updatable_parameter
 
 
 class WeakLensingSystematic(SourceGalaxySystematic[WeakLensingArgs]):
@@ -429,295 +397,3 @@ class HMAlignmentSystematic(WeakLensingSystematic):
         return replace(
             tracer_arg, has_hm=True, ia_a_1h=self.ia_a_1h, ia_a_2h=self.ia_a_2h
         )
-
-
-class WeakLensing(SourceGalaxy[WeakLensingArgs]):
-    """Source class for weak lensing."""
-
-    def __init__(
-        self,
-        *,
-        sacc_tracer: str,
-        scale: float = 1.0,
-        systematics: None | Sequence[SourceGalaxySystematic[WeakLensingArgs]] = None,
-    ):
-        """Initialize the WeakLensing object.
-
-        :param sacc_tracer: the name of the tracer in the SACC file. This is used
-            as a prefix for its parameters.
-        :param scale: the scale of the source. This is used to scale the shear
-            power spectrum.
-        :param systematics: a list of WeakLensingSystematic objects to apply to
-            this source.
-
-        """
-        super().__init__(sacc_tracer=sacc_tracer, systematics=systematics)
-
-        self.sacc_tracer = sacc_tracer
-        self.scale = scale
-        self.current_tracer_args: None | WeakLensingArgs = None
-        self.tracer_args: WeakLensingArgs
-
-    @classmethod
-    def create_ready(
-        cls,
-        inferred_zdist: InferredGalaxyZDist,
-        systematics: None | list[SourceGalaxySystematic[WeakLensingArgs]] = None,
-    ) -> WeakLensing:
-        """Create a WeakLensing object with the given tracer name and scale."""
-        obj = cls(sacc_tracer=inferred_zdist.bin_name, systematics=systematics)
-        # pylint: disable=unexpected-keyword-arg
-        obj.tracer_args = WeakLensingArgs(
-            scale=obj.scale, z=inferred_zdist.z, dndz=inferred_zdist.dndz, ia_bias=None
-        )
-        # pylint: enable=unexpected-keyword-arg
-
-        return obj
-
-    @final
-    def _update_source(self, params: ParamsMap):
-        """Implementation of Source interface `_update_source`.
-
-        This updates all the contained systematics.
-        """
-        self.systematics.update(params)
-
-    def _read(self, sacc_data: sacc.Sacc) -> None:
-        """Read the data for this source from the SACC file.
-
-        This sets self.tracer_args, based on the data in `sacc_data` associated with
-        this object's `sacc_tracer` name.
-        """
-        # pylint: disable=unexpected-keyword-arg
-        self.tracer_args = WeakLensingArgs(
-            scale=self.scale, z=np.array([]), dndz=np.array([]), ia_bias=None
-        )
-        # pylint: enable=unexpected-keyword-arg
-
-        super()._read(sacc_data)
-
-    def create_tracers(self, tools: ModelingTools):
-        """Render a source by applying systematics."""
-        ccl_cosmo = tools.get_ccl_cosmology()
-        tracer_args = self.tracer_args
-
-        assert self.systematics is not None
-        for systematic in self.systematics:
-            tracer_args = systematic.apply(tools, tracer_args)
-
-        ccl_wl_tracer = pyccl.WeakLensingTracer(
-            ccl_cosmo,
-            dndz=(tracer_args.z, tracer_args.dndz),
-            ia_bias=tracer_args.ia_bias,
-        )
-        tracers = [Tracer(ccl_wl_tracer, tracer_name="shear", field=tracer_args.field)]
-
-        if tracer_args.has_pt:
-            ia_pt_tracer = pyccl.nl_pt.PTIntrinsicAlignmentTracer(
-                c1=tracer_args.ia_pt_c_1,
-                cdelta=tracer_args.ia_pt_c_d,
-                c2=tracer_args.ia_pt_c_2,
-            )
-
-            ccl_wl_dummy_tracer = pyccl.WeakLensingTracer(
-                ccl_cosmo,
-                has_shear=False,
-                use_A_ia=False,
-                dndz=(tracer_args.z, tracer_args.dndz),
-                ia_bias=(tracer_args.z, np.ones_like(tracer_args.z)),
-            )
-            ia_tracer = Tracer(
-                ccl_wl_dummy_tracer, tracer_name="intrinsic_pt", pt_tracer=ia_pt_tracer
-            )
-            tracers.append(ia_tracer)
-
-        if tracer_args.has_hm:
-            hmc = tools.get_hm_calculator()
-            cM = tools.get_cM_relation()
-            halo_profile = pyccl.halos.SatelliteShearHOD(
-                mass_def=hmc.mass_def, concentration=cM, a1h=tracer_args.ia_a_1h
-            )
-            ccl_wl_dummy_tracer = pyccl.WeakLensingTracer(
-                ccl_cosmo,
-                has_shear=False,
-                use_A_ia=False,
-                dndz=(tracer_args.z, tracer_args.dndz),
-                ia_bias=(tracer_args.z, np.ones_like(tracer_args.z)),
-            )
-            ia_tracer = Tracer(
-                ccl_wl_dummy_tracer,
-                tracer_name="intrinsic_alignment_hm",
-                halo_profile=halo_profile,
-            )
-            # TODO: redesign this so that we are not adding a new
-            # attribute to a pyccl class.
-            halo_profile.ia_a_2h = (
-                tracer_args.ia_a_2h
-            )  # Attach the 2-halo amplitude here.
-            tracers.append(ia_tracer)
-
-        self.current_tracer_args = tracer_args
-
-        return tracers, tracer_args
-
-    def get_scale(self):
-        """Returns the scales for this Source."""
-        assert self.current_tracer_args
-        return self.current_tracer_args.scale
-
-
-class MultiplicativeShearBiasFactory(BaseModel):
-    """Factory class for MultiplicativeShearBias objects."""
-
-    model_config = ConfigDict(extra="forbid", frozen=True)
-
-    type: Annotated[
-        Literal["MultiplicativeShearBiasFactory"],
-        Field(description="The type of the systematic."),
-    ] = "MultiplicativeShearBiasFactory"
-
-    def create(self, bin_name: str) -> MultiplicativeShearBias:
-        """Create a MultiplicativeShearBias object.
-
-        :param inferred_zdist: The inferred galaxy redshift distribution for
-            the created MultiplicativeShearBias object.
-        :return: The created MultiplicativeShearBias object.
-        """
-        return MultiplicativeShearBias(bin_name)
-
-    def create_global(self) -> MultiplicativeShearBias:
-        """Create a MultiplicativeShearBias object.
-
-        :return: The created MultiplicativeShearBias object.
-        """
-        raise ValueError("MultiplicativeShearBias cannot be global")
-
-
-class LinearAlignmentSystematicFactory(BaseModel):
-    """Factory class for LinearAlignmentSystematic objects."""
-
-    model_config = ConfigDict(extra="forbid", frozen=True)
-
-    type: Annotated[
-        Literal["LinearAlignmentSystematicFactory"],
-        Field(description="The type of the systematic."),
-    ] = "LinearAlignmentSystematicFactory"
-
-    alphag: None | float = 1.0
-
-    def create(self, bin_name: str) -> LinearAlignmentSystematic:
-        """Create a LinearAlignmentSystematic object.
-
-        :param inferred_zdist: The inferred galaxy redshift distribution for
-            the created LinearAlignmentSystematic object.
-        :return: The created LinearAlignmentSystematic object.
-        """
-        return LinearAlignmentSystematic(bin_name)
-
-    def create_global(self) -> LinearAlignmentSystematic:
-        """Create a LinearAlignmentSystematic object.
-
-        :return: The created LinearAlignmentSystematic object.
-        """
-        return LinearAlignmentSystematic(sacc_tracer=None, alphag=self.alphag)
-
-
-class TattAlignmentSystematicFactory(BaseModel):
-    """Factory class for TattAlignmentSystematic objects."""
-
-    model_config = ConfigDict(extra="forbid", frozen=True)
-
-    type: Annotated[
-        Literal["TattAlignmentSystematicFactory"],
-        Field(description="The type of the systematic."),
-    ] = "TattAlignmentSystematicFactory"
-    include_z_dependence: bool = False
-
-    def create(self, bin_name: str) -> TattAlignmentSystematic:
-        """Create a TattAlignmentSystematic object.
-
-        :param inferred_zdist: The inferred galaxy redshift distribution for
-            the created TattAlignmentSystematic object.
-        :return: The created TattAlignmentSystematic object.
-        """
-        return TattAlignmentSystematic(bin_name, self.include_z_dependence)
-
-    def create_global(self) -> TattAlignmentSystematic:
-        """Create a TattAlignmentSystematic object.
-
-        :return: The created TattAlignmentSystematic object.
-        """
-        return TattAlignmentSystematic(None, self.include_z_dependence)
-
-
-WeakLensingSystematicFactory = Annotated[
-    PhotoZShiftFactory
-    | PhotoZShiftandStretchFactory
-    | MultiplicativeShearBiasFactory
-    | LinearAlignmentSystematicFactory
-    | TattAlignmentSystematicFactory,
-    Field(discriminator="type", union_mode="left_to_right"),
-]
-
-
-class WeakLensingFactory(BaseModel):
-    """Factory class for WeakLensing objects."""
-
-    model_config = ConfigDict(extra="forbid", frozen=True)
-
-    _cache: dict[int, WeakLensing] = PrivateAttr()
-    _global_systematics_instances: Sequence[SourceGalaxySystematic[WeakLensingArgs]] = (
-        PrivateAttr()
-    )
-
-    type_source: TypeSource = TypeSource.DEFAULT
-    per_bin_systematics: Sequence[WeakLensingSystematicFactory] = Field(
-        default_factory=list
-    )
-    global_systematics: Sequence[WeakLensingSystematicFactory] = Field(
-        default_factory=list
-    )
-
-    def model_post_init(self, _, /) -> None:
-        """Initialize the WeakLensingFactory object."""
-        self._cache: dict[int, WeakLensing] = {}
-        self._global_systematics_instances = [
-            wl_systematic_factory.create_global()
-            for wl_systematic_factory in self.global_systematics
-        ]
-
-    def create(self, inferred_zdist: InferredGalaxyZDist) -> WeakLensing:
-        """Create a WeakLensing object with the given tracer name and scale."""
-        inferred_zdist_id = id(inferred_zdist)
-        if inferred_zdist_id in self._cache:
-            return self._cache[inferred_zdist_id]
-
-        systematics: list[SourceGalaxySystematic[WeakLensingArgs]] = [
-            systematic_factory.create(inferred_zdist.bin_name)
-            for systematic_factory in self.per_bin_systematics
-        ]
-        systematics.extend(self._global_systematics_instances)
-
-        wl = WeakLensing.create_ready(inferred_zdist, systematics)
-        self._cache[inferred_zdist_id] = wl
-
-        return wl
-
-    def create_from_metadata_only(
-        self,
-        sacc_tracer: str,
-    ) -> WeakLensing:
-        """Create an WeakLensing object with the given tracer name and scale."""
-        sacc_tracer_id = hash(sacc_tracer)  # Improve this
-        if sacc_tracer_id in self._cache:
-            return self._cache[sacc_tracer_id]
-        systematics: list[SourceGalaxySystematic[WeakLensingArgs]] = [
-            systematic_factory.create(sacc_tracer)
-            for systematic_factory in self.per_bin_systematics
-        ]
-        systematics.extend(self._global_systematics_instances)
-
-        wl = WeakLensing(sacc_tracer=sacc_tracer, systematics=systematics)
-        self._cache[sacc_tracer_id] = wl
-
-        return wl
